Change asks and bids type to common.PriceLevel.
