+++ v2/futures/depth_service.go
@@ -2,7 +2,8 @@ package futures
 
 import (
 	"context"
-	"strconv"
+
+	"github.com/adshao/go-binance/v2/common"
 )
 
 // DepthService show depth info
@@ -72,30 +73,8 @@ type DepthResponse struct {
 	Asks         []Ask `json:"asks"`
 }
 
-// PriceLevel is a common structure for bids and asks in the
-// order book.
-type PriceLevel struct {
-	Price    string
-	Quantity string
-}
-
-// Parse parses this PriceLevel's Price and Quantity and
-// returns them both.  It also returns an error if either
-// fails to parse.
-func (p *PriceLevel) Parse() (float64, float64, error) {
-	price, err := strconv.ParseFloat(p.Price, 64)
-	if err != nil {
-		return 0, 0, err
-	}
-	quantity, err := strconv.ParseFloat(p.Quantity, 64)
-	if err != nil {
-		return price, 0, err
-	}
-	return price, quantity, nil
-}
-
 // Ask is a type alias for PriceLevel.
-type Ask = PriceLevel
+type Ask = common.PriceLevel
 
-// Bid is a tpye alias for PriceLevel.
-type Bid = PriceLevel
+// Bid is a type alias for PriceLevel.
+type Bid = common.PriceLevel
